// models/device.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bandwidth
 * Bandwidth\'s Communication APIs
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: letstalk@bandwidth.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DeviceStatusEnum } from './device-status-enum';

export interface Device {
    /**
     * The unique ID of the device.
     */
    'deviceId': string;
    /**
     * The name of the device.
     */
    'deviceName'?: string;
    'status': DeviceStatusEnum;
    /**
     * The time the device was created. In ISO-8601 format.
     */
    'creationTimestamp': string;
}